+++ src/frontend/js/components/Form/Form/index.tsx
@@ -1,0 +1,31 @@
+import c from 'classnames';
+import { PropsWithChildren } from 'react';
+
+interface FormProps extends PropsWithChildren {
+  onSubmit?: (e?: React.BaseSyntheticEvent<object, any, any> | undefined) => Promise<void>;
+  noValidate?: boolean;
+  className?: string;
+  name?: string;
+}
+
+const Form = ({ children, onSubmit, className, name, noValidate = true }: FormProps) => {
+  return (
+    <form className={c('form', className)} name={name} onSubmit={onSubmit} noValidate={noValidate}>
+      {children}
+    </form>
+  );
+};
+
+Form.Row = ({ children, className }: PropsWithChildren<{ className?: string }>) => {
+  return <div className={c('form-row', className)}>{children}</div>;
+};
+
+Form.RowButtonContainer = ({ children, className }: PropsWithChildren<{ className?: string }>) => {
+  return <div className={c('form-row-button-container', className)}>{children}</div>;
+};
+
+Form.Footer = ({ children, className }: PropsWithChildren<{ className?: string }>) => {
+  return <div className={c('form-footer', className)}>{children}</div>;
+};
+
+export default Form;